+++ OggConverter/src/Music/Downloader.cs
@@ -35,7 +35,7 @@ namespace OggConverter
         /// <param name="folder">Radio or CD</param>
         /// <param name="limit">Radio = 99. CD = 15</param>
         /// <returns></returns>
-        public static async Task DownloadFile(string url, string mscPath, string folder, int limit)
+        public static async Task DownloadFile(string url, string mscPath, string folder, int limit, string forcedName = null)
         {
             if (!File.Exists("youtube-dl.exe"))
             {
@@ -94,7 +94,7 @@ namespace OggConverter
             await Task.Run(() => process.WaitForExit());
 
             Form1.instance.Log += "\nConverting...";
-            await Converter.ConvertFile($"{Directory.GetCurrentDirectory()}\\download.aac", mscPath, folder, limit);
+            await Converter.ConvertFile($"{Directory.GetCurrentDirectory()}\\download.aac", mscPath, folder, limit, forcedName);
 
             File.Delete("download.aac");
             IsBusy = false;